LMSA-9452 - more use frontend as paramter vs admin mode

+++ src/main/java/edu/iu/uits/lms/siterequest/controller/admin/SiteRequestAdminController.java
@@ -38,6 +38,7 @@ import edu.iu.uits.lms.canvas.services.AccountService;
 import edu.iu.uits.lms.lti.LTIConstants;
 import edu.iu.uits.lms.lti.controller.OidcTokenAwareController;
 import edu.iu.uits.lms.lti.service.OidcTokenUtils;
+import edu.iu.uits.lms.siterequest.controller.SiteRequestController;
 import edu.iu.uits.lms.siterequest.model.SiteRequestAccountOmit;
 import edu.iu.uits.lms.siterequest.repository.SiteRequestAccountOmitRepository;
 import edu.iu.uits.lms.siterequest.service.SiteRequestOmitAccountService;
@@ -66,6 +67,13 @@ public class SiteRequestAdminController extends OidcTokenAwareController {
 
     @RequestMapping("/launch")
     public String adminMain(Model model) {
+        OidcAuthenticationToken token = getTokenWithoutContext();
+        OidcTokenUtils oidcTokenUtils = new OidcTokenUtils(token);
+
+        if (oidcTokenUtils.getCustomValue(SiteRequestController.IS_FRONTEND_MODE) != null) {
+            return "siterequest_error";
+        }
+
         List<SiteRequestAccountOmit> siteRequestAccountOmits = siteRequestAccountOmitRepository.findAll();
 
         model.addAttribute("omitAccounts", siteRequestAccountOmits);
@@ -75,6 +83,13 @@ public class SiteRequestAdminController extends OidcTokenAwareController {
 
     @RequestMapping("/{omitAccountId}/edit")
     public String adminOmitAccountEdit(@PathVariable("omitAccountId") String omitAccountId, Model model) {
+        OidcAuthenticationToken token = getTokenWithoutContext();
+        OidcTokenUtils oidcTokenUtils = new OidcTokenUtils(token);
+
+        if (oidcTokenUtils.getCustomValue(SiteRequestController.IS_FRONTEND_MODE) != null) {
+            return "siterequest_error";
+        }
+
         SiteRequestAccountOmit siteRequestAccountOmit = siteRequestAccountOmitRepository.findById(Long.parseLong(omitAccountId)).orElse(null);
 
         if (siteRequestAccountOmit != null) {
@@ -86,6 +101,13 @@ public class SiteRequestAdminController extends OidcTokenAwareController {
 
     @RequestMapping("/update")
     public String adminOmitAccountUpdate(@RequestParam("action") String action, SiteRequestAccountOmit siteRequestAccountOmit, Model model) {
+        OidcAuthenticationToken token = getTokenWithoutContext();
+        OidcTokenUtils oidcTokenUtils = new OidcTokenUtils(token);
+
+        if (oidcTokenUtils.getCustomValue(SiteRequestController.IS_FRONTEND_MODE) != null) {
+            return "siterequest_error";
+        }
+
         if ("submit".equalsIgnoreCase(action)) {
             if (siteRequestAccountOmit.getNote() != null) {
                 final int noteLength = siteRequestAccountOmit.getNote().length();
@@ -106,6 +128,10 @@ public class SiteRequestAdminController extends OidcTokenAwareController {
         OidcAuthenticationToken token = getTokenWithoutContext();
         OidcTokenUtils oidcTokenUtils = new OidcTokenUtils(token);
 
+        if (oidcTokenUtils.getCustomValue(SiteRequestController.IS_FRONTEND_MODE) != null) {
+            return "siterequest_error";
+        }
+
         model.addAttribute("create", true);
 
         SiteRequestAccountOmit siteRequestAccountOmit = new SiteRequestAccountOmit();
@@ -117,6 +143,13 @@ public class SiteRequestAdminController extends OidcTokenAwareController {
 
     @RequestMapping("/submit")
     public String adminOmitAccountCreate(@RequestParam("action") String action, SiteRequestAccountOmit siteRequestAccountOmit, Model model) {
+        OidcAuthenticationToken token = getTokenWithoutContext();
+        OidcTokenUtils oidcTokenUtils = new OidcTokenUtils(token);
+
+        if (oidcTokenUtils.getCustomValue(SiteRequestController.IS_FRONTEND_MODE) != null) {
+            return "siterequest_error";
+        }
+
         if ("submit".equalsIgnoreCase(action)) {
             if (siteRequestAccountOmit.getNote() != null) {
                 final int noteLength = siteRequestAccountOmit.getNote().length();
@@ -142,6 +175,13 @@ public class SiteRequestAdminController extends OidcTokenAwareController {
 
     @RequestMapping("/{omitAccountId}/delete")
     public String adminOmitAccountDelete(@PathVariable("omitAccountId") String omitAccountId, Model model) {
+        OidcAuthenticationToken token = getTokenWithoutContext();
+        OidcTokenUtils oidcTokenUtils = new OidcTokenUtils(token);
+
+        if (oidcTokenUtils.getCustomValue(SiteRequestController.IS_FRONTEND_MODE) != null) {
+            return "siterequest_error";
+        }
+
         siteRequestAccountOmitRepository.deleteById(Long.parseLong(omitAccountId));
 
         return adminMain(model);
